backend/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "concept-visualizer-backend"
version = "0.1.0"
description = "Backend API for the Concept Visualizer project"
readme = "README.md"
requires-python = ">=3.11"
authors = [
    {name = "Concept Visualizer Team"},
]
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn>=0.23.2",
    "pydantic>=2.4.2",
    "httpx>=0.25.0",
    "python-dotenv>=1.0.0",
    "python-multipart>=0.0.6",
    "pydantic-settings>=2.8.1",
    "jigsawstack>=0.1.31",
    "supabase>=1.0.3",
    "pillow>=10.0.0",
    "opencv-python>=4.8.0",
    "vtracer>=0.0.11",
    "slowapi>=0.1.9",
    "redis>=5.0.0",
    "upstash-redis>=1.0.0",
    "python-jose>=3.3.0",
    "pytest-asyncio>=0.26.0",
    "pre-commit>=4.2.0",
    "types-python-dateutil>=2.9.0.20241206",
    "qrcode>=7.4.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.9.1",
    "isort>=5.12.0",
    "mypy>=1.5.1",
    "types-requests>=2.31.0.2",
    "flake8>=6.1.0",
    "flake8-docstrings>=1.7.0",
]

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.black]
line-length = 200
target-version = ["py311"]

[tool.isort]
profile = "black"
line_length = 200

[tool.mypy]
python_version = "3.11"
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
strict_optional = true
warn_redundant_casts = true
warn_return_any = true
warn_unused_ignores = true
explicit_package_bases = true
mypy_path = "."
namespace_packages = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
